+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Renamed `Lightstreamer::Subscription#add_data_callback` to `Lightstreamer::Subscription#on_data`
 - Replaced `Lightstreamer::Subscription#remove_data_callback` with `Lightstreamer::Subscription#clear_callbacks`
 - Overflow messages can now be processed by the application using `Lightstreamer::Subscription#on_callback`
+- Renamed `Lightstreamer::Error` class to `Lightstreamer::LightstreamerError`
 - Improved documentation
 
 ### 0.4 — July 25, 2016